Re organize
Just re organized everything still need to add stuff and fill the folders.
