+++ client/ToDoApp/src/main.tsx
@@ -2,19 +2,19 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 //Routing imports
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { TaskProvider } from "./TaskProvider";
+import { TaskProvider } from "./Components/TaskProvider";
 //Component imports
-import ToDoCreater from "./Tasks/ToDoCreater";
-import Container from "./Tasks/TodoContainer";
-import Nav from "./Nav/Navigator";
-import Login from "./login/LoginSign/Login";
-import Groups from "./Folders/groups";
+import ToDoCreater from "./Pages/Tasks/ToDoCreater";
+import Container from "./Pages/Tasks/TodoContainer";
+import Nav from "./Components/Nav/Navigator";
+import Login from "./Pages/LoginSign/Login";
+import Groups from "./Pages/Folders/groups";
 //CSS imports
-import "./login/LoginSign/NavLogin.css";
-import "./Folders/Check-Folder.css";
-import "./Tasks/Task.css";
-import "./Tasks/Creator.css";
-import "./Nav/Nav.css";
+import "./Pages/LoginSign/NavLogin.css";
+import "./Pages/Folders/Check-Folder.css";
+import "./Pages/Tasks/Task.css";
+import "./Pages/Tasks/Creator.css";
+import "./Components/Nav/Nav.css";
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <TaskProvider>
