--- client/ToDoApp/tests/ToDoCreater.test.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { render, screen, fireEvent } from "@testing-library/react";
-import "@testing-library/jest-dom/extend-expect";
-import ToDoCreater from "./ToDoCreater";
-
-describe("ToDoCreater Component", () => {
-  test("renders the ToDoCreater component", () => {});
-});
